updated the structure to include maven & utilize it for our junit 5 dependency
