--- src/Airline/AirlineController.java
@@ -1,4 +1,0 @@
-package Airline;
-
-public class AirlineController {
-}
