--- src/Airport/Airport.java
@@ -1,4 +1,0 @@
-package Airport;
-
-public class Airport {
-}
