--- src/Airport/AirportService.java
@@ -1,4 +1,0 @@
-package Airport;
-
-public class AirportService {
-}
