--- src/Member/Member.java
@@ -1,4 +1,0 @@
-package Member;
-
-public class Member {
-}
