--- src/Member/MemberController.java
@@ -1,4 +1,0 @@
-package Member;
-
-public class MemberController {
-}
