--- src/Plane/Plane.java
@@ -1,4 +1,0 @@
-package Plane;
-
-public class Plane {
-}
