+++ src/main/java/com/revature/ams/Airline/Airline.java
@@ -1,0 +1,4 @@
+package com.revature.ams.Airline;
+
+public class Airline {
+}
