+++ src/main/java/com/revature/ams/Airport/AirportService.java
@@ -1,0 +1,4 @@
+package com.revature.ams.Airport;
+
+public class AirportService {
+}
