+++ src/main/java/com/revature/ams/Member/MemberController.java
@@ -1,0 +1,4 @@
+package com.revature.ams.Member;
+
+public class MemberController {
+}
