+++ src/main/java/com/revature/ams/Plane/PlaneController.java
@@ -1,0 +1,4 @@
+package com.revature.ams.Plane;
+
+public class PlaneController {
+}
