+++ src/main/java/com/revature/ams/util/exceptions/InvalidInputException.java
@@ -1,0 +1,8 @@
+package com.revature.ams.util.exceptions;
+
+// OOP - Inheritence -
+public class InvalidInputException extends Exception{
+    public InvalidInputException(String message){
+        super(message); // the constructor from the inherited class above
+    }
+}
