--- src/util/exceptions/InvalidInputException.java
@@ -1,8 +1,0 @@
-package util.exceptions;
-
-// OOP - Inheritence -
-public class InvalidInputException extends Exception{
-    public InvalidInputException(String message){
-        super(message); // the constructor from the inherited class above
-    }
-}
